requirements.txt
adapter-transformers==4.24.0
sklearn==1.1.2
torch==1.12.1
torchvision==0.13.1
